moved client to own component, matching prev file structure-control comp. and styling still to add, but functional
